Thanks to GerHobbelt for helping me track this one done.
The title element needs to be hidden, but we can't use the hide/show
methods here, because they are overloaded to look for an instance
of windows/panels.  In this case the title element is has the window
instance attached to it.  So hide hides the window not the title.